+++ vcfa/config_test.go
@@ -22,13 +22,7 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/kr/pretty"
-
-	semver "github.com/hashicorp/go-version"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/vmware/go-vcloud-director/v3/govcd"
 	"github.com/vmware/go-vcloud-director/v3/util"
 )
@@ -81,8 +75,8 @@ type TestConfig struct {
 
 		// Versions are filled by the VCFA provisioning task, and allow tests to
 		// check for compatibility without using an extra connection
-		VcdVersion string `json:"vcdVersion,omitempty"`
-		ApiVersion string `json:"apiVersion,omitempty"`
+		TmVersion    string `json:"tmVersion,omitempty"`
+		TmApiVersion string `json:"tmApiVersion,omitempty"`
 
 		// UseSamlAdfs specifies if SAML auth is used for authenticating vCD instead of local login.
 		// The above `User` and `Password` will be used to authenticate against ADFS IdP when true.
@@ -131,114 +125,14 @@ type TestConfig struct {
 		VcenterSupervisorZone string `json:"vcenterSupervisorZone"`
 	} `json:"tm,omitempty"`
 	VCD struct {
-		Org         string `json:"org"`
-		Vdc         string `json:"vdc"`
-		ProviderVdc struct {
-			Name            string `json:"name"`
-			NetworkPool     string `json:"networkPool"`
-			StorageProfile  string `json:"storageProfile"`
-			StorageProfile2 string `json:"storageProfile2"`
-		} `json:"providerVdc"`
-		NsxtProviderVdc struct {
-			Name                   string `json:"name"`
-			StorageProfile         string `json:"storageProfile"`
-			StorageProfile2        string `json:"storageProfile2"`
-			NetworkPool            string `json:"networkPool"`
-			PlacementPolicyVmGroup string `json:"placementPolicyVmGroup"` // Name of the VM group to create VM Placement Policies
-		} `json:"nsxtProviderVdc"`
-		Catalog struct {
-			Name                      string `json:"name,omitempty"`
-			CatalogItem               string `json:"catalogItem,omitempty"`
-			CatalogItemWithEfiSupport string `json:"catalogItemWithEfiSupport,omitempty"`
-			CatalogItemWithMultiVms   string `json:"catalogItemWithMultiVms,omitempty"`
-			VmName1InMultiVmItem      string `json:"vmName1InMultiVmItem,omitempty"`
-			VmName2InMultiVmItem      string `json:"VmName2InMultiVmItem,omitempty"`
-			NsxtBackedCatalogName     string `json:"nsxtBackedCatalogName,omitempty"`
-			NsxtCatalogAddonDse       string `json:"nsxtCatalogAddonDse,omitempty"`
-			NsxtCatalogItem           string `json:"nsxtCatalogItem,omitempty"`
-			VSphereSubscribedCatalog  string `json:"vSphereSubscribedCatalog,omitempty"`
-		} `json:"catalog"`
+		Org string `json:"org"`
 	} `json:"vcd"`
-	Networking struct {
-		ExternalIp                   string `json:"externalIp,omitempty"`
-		InternalIp                   string `json:"internalIp,omitempty"`
-		EdgeGateway                  string `json:"edgeGateway,omitempty"`
-		SharedSecret                 string `json:"sharedSecret"`
-		Vcenter                      string `json:"vcenter,omitempty"`
-		ExternalNetwork              string `json:"externalNetwork,omitempty"`
-		ExternalNetworkPortGroup     string `json:"externalNetworkPortGroup,omitempty"`
-		ExternalNetworkPortGroupType string `json:"externalNetworkPortGroupType,omitempty"`
-		LdapServer                   string `json:"ldapServer,omitempty"`
-		OidcServer                   struct {
-			Url               string `json:"url,omitempty"`
-			WellKnownEndpoint string `json:"wellKnownEndpoint,omitempty"`
-		} `json:"oidcServer"`
-		Local struct {
-			LocalIp            string `json:"localIp"`
-			LocalSubnetGateway string `json:"localSubnetGw"`
-		} `json:"local"`
-		Peer struct {
-			PeerIp            string `json:"peerIp"`
-			PeerSubnetGateway string `json:"peerSubnetGw"`
-		} `json:"peer"`
-	} `json:"networking"`
-	Nsxt struct {
-		Manager                   string `json:"manager"`
-		Tier0router               string `json:"tier0router"`
-		Tier0routerInterface      string `json:"tier0routerInterface"`
-		Tier0routerVrf            string `json:"tier0routervrf"`
-		GatewayQosProfile         string `json:"gatewayQosProfile"`
-		NsxtDvpg                  string `json:"nsxtDvpg"`
-		Vdc                       string `json:"vdc"`
-		ExternalNetwork           string `json:"externalNetwork"`
-		EdgeGateway               string `json:"edgeGateway"`
-		VdcGroup                  string `json:"vdcGroup"`
-		VdcGroupEdgeGateway       string `json:"vdcGroupEdgeGateway"`
-		NsxtImportSegment         string `json:"nsxtImportSegment"`
-		NsxtImportSegment2        string `json:"nsxtImportSegment2"`
-		NsxtEdgeCluster           string `json:"nsxtEdgeCluster"`
-		NsxtAlbControllerUrl      string `json:"nsxtAlbControllerUrl"`
-		NsxtAlbControllerUser     string `json:"nsxtAlbControllerUser"`
-		NsxtAlbControllerPassword string `json:"nsxtAlbControllerPassword"`
-		NsxtAlbImportableCloud    string `json:"nsxtAlbImportableCloud"`
-		NsxtAlbServiceEngineGroup string `json:"nsxtAlbServiceEngineGroup"`
-		RoutedNetwork             string `json:"routedNetwork"`
-		IsolatedNetwork           string `json:"isolatedNetwork"`
-		DirectNetwork             string `json:"directNetwork"`
-		IpDiscoveryProfile        string `json:"ipDiscoveryProfile"`
-		MacDiscoveryProfile       string `json:"macDiscoveryProfile"`
-		SpoofGuardProfile         string `json:"spoofGuardProfile"`
-		QosProfile                string `json:"qosProfile"`
-		SegmentSecurityProfile    string `json:"segmentSecurityProfile"`
-	} `json:"nsxt"`
-	VSphere struct {
-		ResourcePoolForVcd1 string `json:"resourcePoolForVcd1,omitempty"`
-		ResourcePoolForVcd2 string `json:"resourcePoolForVcd2,omitempty"`
-	} `json:"vsphere,omitempty"`
 	Logging struct {
 		Enabled         bool   `json:"enabled,omitempty"`
 		LogFileName     string `json:"logFileName,omitempty"`
 		LogHttpRequest  bool   `json:"logHttpRequest,omitempty"`
 		LogHttpResponse bool   `json:"logHttpResponse,omitempty"`
 	} `json:"logging"`
-	Ova struct {
-		OvaPath             string `json:"ovaPath,omitempty"`
-		OvfUrl              string `json:"ovfUrl,omitempty"`
-		UploadPieceSize     int64  `json:"uploadPieceSize,omitempty"`
-		UploadProgress      bool   `json:"uploadProgress,omitempty"`
-		OvaTestFileName     string `json:"ovaTestFileName,omitempty"`
-		OvaDownloadUrl      string `json:"ovaDownloadUrl,omitempty"`
-		Preserve            bool   `json:"preserve,omitempty"`
-		OvaVappMultiVmsPath string `json:"ovaVappMultiVmsPath,omitempty"`
-	} `json:"ova"`
-	Media struct {
-		MediaPath           string `json:"mediaPath,omitempty"`
-		UploadPieceSize     int64  `json:"uploadPieceSize,omitempty"`
-		UploadProgress      bool   `json:"uploadProgress,omitempty"`
-		MediaName           string `json:"mediaName,omitempty"`
-		NsxtBackedMediaName string `json:"nsxtBackedMediaName,omitempty"`
-		UiPluginPath        string `json:"uiPluginPath,omitempty"`
-	} `json:"media"`
 	Certificates struct {
 		Certificate1Path           string `json:"certificate1Path,omitempty"`           // absolute path to pem file
 		Certificate1PrivateKeyPath string `json:"certificate1PrivateKeyPath,omitempty"` // absolute path to private key pem file
@@ -248,55 +142,11 @@ type TestConfig struct {
 		Certificate2Pass           string `json:"certificate2Pass,omitempty"`           // absolute path to pem file
 		RootCertificatePath        string `json:"rootCertificatePath,omitempty"`        // absolute path to pem file
 	} `json:"certificates"`
-	// Data used to create a new environment, in addition to the regular test configuration file
-	TestEnvBuild struct {
-		Gateway                      string `json:"gateway"`                      // Gateway for external network
-		Netmask                      string `json:"netmask"`                      // Netmask for external network
-		ExternalNetworkStartIp       string `json:"externalNetworkStartIp"`       // Start IP for external network
-		ExternalNetworkEndIp         string `json:"externalNetworkEndIp"`         // End IP for external network
-		Dns1                         string `json:"dns1"`                         // DNS 1 for external network
-		Dns2                         string `json:"dns2"`                         // DNS 2 for external network
-		ExternalNetworkPortGroup     string `json:"externalNetworkPortGroup"`     // port group, if different from Networking.ExternalNetworkPortGroup
-		ExternalNetworkPortGroupType string `json:"externalNetworkPortGroupType"` // port group type, if different from Networking.ExternalNetworkPortGroupType
-		RoutedNetwork                string `json:"routedNetwork"`                // optional routed network name to create
-		IsolatedNetwork              string `json:"isolatedNetwork"`              // optional isolated network name to create
-		DirectNetwork                string `json:"directNetwork"`                // optional direct network name to create
-		MediaPath                    string `json:"mediaPath"`                    // Media path, if different from Media.MediaPath
-		MediaName                    string `json:"mediaName"`                    // Media name to create
-		OvaPath                      string `json:"ovaPath"`                      // Ova Path, if different from Ova.OvaPath
-		OrgUser                      string `json:"orgUser"`                      // Org User to be created within the organization
-		OrgUserPassword              string `json:"orgUserPassword"`              // Password for the Org User to be created within the organization
-	} `json:"testEnvBuild"`
 	EnvVariables map[string]string `json:"envVariables,omitempty"`
-	Cse          struct {
-		Version        string `json:"version,omitempty"`
-		StorageProfile string `json:"storageProfile,omitempty"`
-		SolutionsOrg   string `json:"solutionsOrg,omitempty"`
-		TenantOrg      string `json:"tenantOrg,omitempty"`
-		TenantVdc      string `json:"tenantVdc,omitempty"`
-		OvaCatalog     string `json:"ovaCatalog,omitempty"`
-		OvaName        string `json:"ovaName,omitempty"`
-		RoutedNetwork  string `json:"routedNetwork,omitempty"`
-		EdgeGateway    string `json:"edgeGateway,omitempty"`
-	} `json:"cse,omitempty"`
-	SolutionAddOn struct {
-		Org           string `json:"org,omitempty"`
-		Vdc           string `json:"vdc,omitempty"`
-		RoutedNetwork string `json:"routedNetwork,omitempty"`
-		ComputePolicy string `json:"computePolicy,omitempty"`
-		StoragePolicy string `json:"storagePolicy,omitempty"`
-
-		Catalog       string                       `json:"catalog,omitempty"`
-		AddOnImageDse string                       `json:"addonImageDse,omitempty"`
-		DseSolutions  map[string]map[string]string `json:"dseSolutions,omitempty"`
-	} `json:"solutionAddOn,omitempty"`
 }
 
 // names for created resources for all the tests
 var (
-	testSuiteCatalogName    = "TestSuiteCatalog"
-	testSuiteCatalogOVAItem = "TestSuiteOVA"
-
 	// vcfaAddProvider will add the provide section to the template
 	vcfaAddProvider = os.Getenv(envVcfaAddProvider) != ""
 
@@ -361,17 +211,13 @@ var (
 )
 
 const (
-	providerVcfaSystem    = "vcfa"
 	providerVcfaOrg1      = "vcfaorg1"
 	providerVcfaOrg1Alias = "vcfa.org1"
 	providerVcfaOrg2      = "vcfaorg2"
 	providerVcfaOrg2Alias = "vcfa.org2"
-	providerVcfaSystem1   = "vcfasys1"
-	providerVcfaSys1Alias = "vcfa.sys1"
 	providerVcfaSystem2   = "vcfasys2"
 	providerVcfaSys2Alias = "vcfa.sys2"
 
-	customTemplatesDirectory        = "test-templates"
 	testArtifactsDirectory          = "test-artifacts"
 	envVcfaAddProvider              = "VCFA_ADD_PROVIDER"
 	envVcfaSkipTemplateWriting      = "VCFA_SKIP_TEMPLATE_WRITING"
@@ -388,7 +234,7 @@ const (
 # date {{.Timestamp}}
 # file {{.CallerFileName}}
 # VCFA version {{.VcfaVersion}}
-# API version {{.ApiVersion}}
+# API version {{.tmApiVersion}}
 
 provider "vcfa" {
   user                 = "{{.PrUser}}"
@@ -400,12 +246,10 @@ provider "vcfa" {
   url                  = "{{.PrUrl}}"
   sysorg               = "{{.PrSysOrg}}"
   org                  = "{{.PrOrg}}"
-  vdc                  = "{{.PrVdc}}"
   allow_unverified_ssl = "{{.AllowInsecure}}"
   max_retry_timeout    = {{.MaxRetryTimeout}}
   logging              = {{.Logging}}
   logging_file         = "{{.LoggingFile}}"
-  {{.IgnoreMetadataBlock}}
 }
 `
 )
@@ -423,10 +267,6 @@ var (
 	testArtifactNames = make(map[string]string)
 )
 
-func testDistributedNetworksEnabled() bool {
-	return testDistributedNetworks || os.Getenv("VCFA_TEST_DISTRIBUTED_NETWORK") != ""
-}
-
 // usingSysAdmin returns true if the current configuration uses a system administrator for connections
 func usingSysAdmin() bool {
 	return strings.ToLower(testConfig.Provider.SysOrg) == "system"
@@ -436,12 +276,6 @@ func usingSysAdmin() bool {
 func skipIfNotSysAdmin(t *testing.T) {
 	if !usingSysAdmin() {
 		t.Skip(t.Name() + " requires system admin privileges")
-	}
-}
-
-func skipIfNotTm(t *testing.T) {
-	if checkVersion(testConfig.Provider.ApiVersion, "< 40.0") {
-		t.Skip(t.Name() + " requires 'tm'")
 	}
 }
 
@@ -530,10 +364,6 @@ func templateFill(tmpl string, inputData StringMap) string {
 		data["PrUrl"] = testConfig.Provider.Url
 		data["PrSysOrg"] = testConfig.Provider.SysOrg
 		data["PrOrg"] = testConfig.VCD.Org
-		vdcName, found := data["PrVdc"]
-		if !found || vdcName == "" {
-			data["PrVdc"] = testConfig.Nsxt.Vdc
-		}
 		data["AllowInsecure"] = testConfig.Provider.AllowInsecure
 		data["MaxRetryTimeout"] = testConfig.Provider.MaxRetryTimeout
 		data["VersionRequired"] = currentProviderVersion
@@ -542,9 +372,6 @@ func templateFill(tmpl string, inputData StringMap) string {
 			data["LoggingFile"] = testConfig.Logging.LogFileName
 		} else {
 			data["LoggingFile"] = util.ApiLogFileName
-		}
-		if _, found = data["IgnoreMetadataBlock"]; !found {
-			data["IgnoreMetadataBlock"] = ""
 		}
 
 		// Pick correct auth_type
@@ -572,8 +399,8 @@ func templateFill(tmpl string, inputData StringMap) string {
 		data["CallerFileName"] = callerFileName
 	}
 	data["Timestamp"] = time.Now().Format("2006-01-02 15:04")
-	data["VcdVersion"] = testConfig.Provider.VcdVersion
-	data["ApiVersion"] = testConfig.Provider.ApiVersion
+	data["tmVersion"] = testConfig.Provider.TmVersion
+	data["tmApiVersion"] = testConfig.Provider.TmApiVersion
 
 	// Creates a template. The template gets the same name of the calling function, to generate a better
 	// error message in case of failure
@@ -732,17 +559,6 @@ func getConfigStruct(config string) TestConfig {
 		configStruct.Provider.SysOrg = configStruct.VCD.Org
 	}
 
-	if vcfaTestOrgUser {
-		user := configStruct.TestEnvBuild.OrgUser
-		password := configStruct.TestEnvBuild.OrgUserPassword
-		if user == "" || password == "" {
-			panic(fmt.Sprintf("%s was enabled, but org user credentials were not found in the configuration file", envVcfaTestOrgUser))
-		}
-		configStruct.Provider.User = user
-		configStruct.Provider.Password = password
-		configStruct.Provider.SysOrg = configStruct.VCD.Org
-		fmt.Println("VCFA_TEST_ORG_USER was enabled. Using Org User credentials from configuration file")
-	}
 	if configStruct.Provider.Token != "" && configStruct.Provider.Password == "" {
 		configStruct.Provider.Password = "TOKEN"
 	}
@@ -763,7 +579,6 @@ func getConfigStruct(config string) TestConfig {
 	_ = os.Setenv("VCFA_URL", configStruct.Provider.Url)
 	_ = os.Setenv("VCFA_SYS_ORG", configStruct.Provider.SysOrg)
 	_ = os.Setenv("VCFA_ORG", configStruct.VCD.Org)
-	_ = os.Setenv("VCFA_VDC", configStruct.Nsxt.Vdc)
 	if configStruct.Provider.UseVcdConnectionCache {
 		enableConnectionCache = true
 	}
@@ -786,34 +601,6 @@ func getConfigStruct(config string) TestConfig {
 		util.InitLogging()
 	}
 
-	if configStruct.Ova.OvaPath != "" {
-		ovaPath, err := filepath.Abs(configStruct.Ova.OvaPath)
-		if err != nil {
-			panic("error retrieving absolute path for OVA path " + configStruct.Ova.OvaPath)
-		}
-		configStruct.Ova.OvaPath = ovaPath
-	}
-	if configStruct.Media.MediaPath != "" {
-		mediaPath, err := filepath.Abs(configStruct.Media.MediaPath)
-		if err != nil {
-			panic("error retrieving absolute path for Media path " + configStruct.Media.MediaPath)
-		}
-		configStruct.Media.MediaPath = mediaPath
-	}
-	if configStruct.Media.UiPluginPath != "" {
-		uiPluginPath, err := filepath.Abs(configStruct.Media.UiPluginPath)
-		if err != nil {
-			panic("error retrieving absolute path for UI plugin path " + configStruct.Media.UiPluginPath)
-		}
-		configStruct.Media.UiPluginPath = uiPluginPath
-	}
-	if configStruct.Ova.OvaVappMultiVmsPath != "" {
-		multiVmOvaPath, err := filepath.Abs(configStruct.Ova.OvaVappMultiVmsPath)
-		if err != nil {
-			panic("error retrieving absolute path for multi OVA path " + configStruct.Ova.OvaVappMultiVmsPath)
-		}
-		configStruct.Ova.OvaVappMultiVmsPath = multiVmOvaPath
-	}
 	if configStruct.Certificates.Certificate1Path != "" {
 		certificatePath1Path, err := filepath.Abs(configStruct.Certificates.Certificate1Path)
 		if err != nil {
@@ -848,14 +635,6 @@ func getConfigStruct(config string) TestConfig {
 			panic("error retrieving absolute path for certificate 2 path " + configStruct.Certificates.Certificate2Path)
 		}
 		configStruct.Certificates.RootCertificatePath = rootCertificatePath2Path
-	}
-
-	// It is needed when we run the binary tests without TEST_ACC
-	if configStruct.VCD.Catalog.Name != "" {
-		testSuiteCatalogName = configStruct.VCD.Catalog.Name
-	}
-	if configStruct.VCD.Catalog.CatalogItem != "" {
-		testSuiteCatalogOVAItem = configStruct.VCD.Catalog.CatalogItem
 	}
 	return configStruct
 }
@@ -983,7 +762,7 @@ func TestMain(m *testing.M) {
 	exitCode := m.Run()
 
 	if numberOfPartitions != 0 {
-		entTestFileName := getTestFileName("end", testConfig.Provider.VcdVersion)
+		entTestFileName := getTestFileName("end", testConfig.Provider.TmVersion)
 		err := os.WriteFile(entTestFileName, []byte(fmt.Sprintf("%d", exitCode)), 0600)
 		if err != nil {
 			fmt.Printf("error writing to file '%s': %s\n", entTestFileName, err)
@@ -1029,186 +808,6 @@ func getTestVCFAFromJson(testConfig TestConfig) (*govcd.VCDClient, error) {
 	return vcdClient, nil
 }
 
-// Used by resources at the top of the hierarchy (such as Org, ExternalNetwork)
-func importStateIdTopHierarchy(objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		return objectName, nil
-	}
-}
-
-// Used by all entities that depend on Org (such as Catalog, OrgUser)
-// If the orgName is empty, it uses the default Org from testConfig
-func importStateIdOrgObject(orgName string, objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if orgName == "" {
-			orgName = testConfig.VCD.Org
-		}
-		if orgName == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return orgName +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// Used by all entities that depend on Org + VDC (such as Vapp, networks, edge gateway)
-func importStateIdOrgVdcObject(objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || testConfig.VCD.Vdc == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			testConfig.VCD.Vdc +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// importStateIdOrgNsxtVdcObject can be used by all entities that depend on Org + NSX-T VDC (such as Vapp, networks,
-// edge gateway) in NSX-T VDC
-func importStateIdOrgNsxtVdcObject(objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || testConfig.Nsxt.Vdc == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			testConfig.Nsxt.Vdc +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// importStateIdOrgNsxtVdcGroupObject can be used by all entities that depend on Org + NSX-T VDC
-// Group (such as Vapp, networks, edge gateway) in NSX-T VDC
-func importStateIdOrgNsxtVdcGroupObject(vdcGroupName, objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			vdcGroupName +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// importCustomObject accepts a path and joins it using ImportSeparator
-func importCustomObject(path []string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		return strings.Join(path, ImportSeparator), nil
-	}
-}
-
-// importStateIdNsxtManagerObject can be used by all entities that depend on NSX-T manager name + objectName
-func importStateIdNsxtManagerObject(objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.Nsxt.Manager == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.Nsxt.Manager + ImportSeparator + objectName, nil
-	}
-}
-
-// Used by all entities that depend on Org + Catalog (such as catalog item, media item)
-func importStateIdOrgCatalogObject(objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || testConfig.VCD.Catalog.Name == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			testConfig.VCD.Catalog.Name +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// Used by all entities that depend on Org + VDC + vApp (such as VM, vapp networks)
-func importStateIdVappObject(vappName, objectName, vdc string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || vappName == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			vdc +
-			ImportSeparator +
-			vappName +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// Used by all entities that depend on Org + VDC + edge gateway (such as FW, LB, NAT)
-func importStateIdEdgeGatewayObject(edgeGatewayName, objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || testConfig.VCD.Vdc == "" || edgeGatewayName == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			testConfig.VCD.Vdc +
-			ImportSeparator +
-			edgeGatewayName +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// importStateIdNsxtEdgeGatewayObject used by all entities that depend on Org + NSX-T VDC + edge gateway (such as FW, NAT, Security Groups)
-func importStateIdNsxtEdgeGatewayObject(edgeGatewayName, objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || testConfig.Nsxt.Vdc == "" || edgeGatewayName == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path for object %s, %s,%s,%s", objectName, testConfig.VCD.Org, testConfig.Nsxt.Vdc, edgeGatewayName)
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			testConfig.Nsxt.Vdc +
-			ImportSeparator +
-			edgeGatewayName +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
-// Used by all entities that depend on Org + VDC + vApp VM (such as VM internal disks)
-func importStateIdVmObject(orgName, vdcName, vappName, vmName, objectIdentifier string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if orgName == "" || vdcName == "" || vappName == "" || vmName == "" || objectIdentifier == "" {
-			return "", fmt.Errorf("missing information to generate import path")
-		}
-		return orgName +
-			ImportSeparator +
-			vdcName +
-			ImportSeparator +
-			vappName +
-			ImportSeparator +
-			vmName +
-			ImportSeparator +
-			objectIdentifier, nil
-	}
-}
-
-// importStateIdNsxtEdgeGatewayObjectUsingVdcGroup used by all entities that depend on Org + NSX-T edge gateway (such as IP Sets, Security Groups)
-func importStateIdNsxtEdgeGatewayObjectUsingVdcGroup(vdcGroupName, edgeGatewayName, objectName string) resource.ImportStateIdFunc {
-	return func(*terraform.State) (string, error) {
-		if testConfig.VCD.Org == "" || vdcGroupName == "" || edgeGatewayName == "" || objectName == "" {
-			return "", fmt.Errorf("missing information to generate import path for object %s", objectName)
-		}
-		return testConfig.VCD.Org +
-			ImportSeparator +
-			vdcGroupName +
-			ImportSeparator +
-			edgeGatewayName +
-			ImportSeparator +
-			objectName, nil
-	}
-}
-
 // setBoolFlag binds a flag to a boolean variable (passed as pointer)
 // it also uses an optional environment variable that, if set, will
 // update the variable before binding it to the flag.
@@ -1241,198 +840,6 @@ func setIntFlag(varPointer *int, name, envVar, help string) {
 	flag.IntVar(varPointer, name, *varPointer, help)
 }
 
-type envHelper struct {
-	vars map[string]string
-}
-
-// newEnvVarHelper helps to initialize
-func newEnvVarHelper() *envHelper {
-	return &envHelper{vars: make(map[string]string)}
-}
-
-// saveVcfaVars checks all env vars with VCFA prefix and saves them in a map
-func (env *envHelper) saveVcfaVars() {
-	for _, envVar := range os.Environ() {
-		if strings.HasPrefix(envVar, "VCFA") {
-			// os.Environ returns a slice of "key=value" strings. The first "=" separated "key" and
-			// "value" therefore we split only first "=" match as env vars may have syntax of
-			// "key=value=else"
-			splitKeyValue := strings.SplitN(envVar, "=", 2)
-			key := splitKeyValue[0]
-			value := splitKeyValue[1]
-			env.vars[key] = value
-		}
-	}
-
-}
-
-// unsetVcfaVars unsets all environment variables with prefix "VCFA"
-func (env *envHelper) unsetVcfaVars() {
-	for keyName := range env.vars {
-		os.Unsetenv(keyName)
-	}
-}
-
-// restoreVcfaVars restores all env variables with prefix "VCFA" stored in parent struct
-func (env *envHelper) restoreVcfaVars() {
-	for keyName, valueName := range env.vars {
-		err := os.Setenv(keyName, valueName)
-		if err != nil {
-			util.Logger.Printf("[ERROR] error setting environment variable %s with value %s", keyName, valueName)
-		}
-	}
-}
-
-// importStateIdViaResource runs the import of a VM affinity rule using the resource ID
-func importStateIdViaResource(resource string) resource.ImportStateIdFunc {
-	return func(s *terraform.State) (string, error) {
-		rs, ok := s.RootModule().Resources[resource]
-		if !ok {
-			return "", fmt.Errorf("resource not found: %s", resource)
-		}
-
-		if rs.Primary.ID == "" {
-			return "", fmt.Errorf("no ID is set for %s resource", resource)
-		}
-
-		importId := testConfig.VCD.Org + "." + testConfig.Nsxt.Vdc + "." + rs.Primary.ID
-		if testConfig.VCD.Org == "" || testConfig.Nsxt.Vdc == "" || rs.Primary.ID == "" {
-			return "", fmt.Errorf("missing information to generate import path: %s", importId)
-		}
-		return importId, nil
-	}
-}
-
-func importStateCatalogIdViaResource(resourceDef string) resource.ImportStateIdFunc {
-	return func(s *terraform.State) (string, error) {
-		rs, ok := s.RootModule().Resources[resourceDef]
-		if !ok {
-			return "", fmt.Errorf("resource not found: %s", resourceDef)
-		}
-
-		if rs.Primary.ID == "" {
-			return "", fmt.Errorf("no ID is set for %s resource", resourceDef)
-		}
-
-		importId := testConfig.VCD.Org + "." + rs.Primary.ID
-		if testConfig.VCD.Org == "" || rs.Primary.ID == "" {
-			return "", fmt.Errorf("missing information to generate import path: %s", importId)
-		}
-		return importId, nil
-	}
-}
-
-// importStateCatalogId returns a catalog ID from a pair of org+catalog names
-func importStateCatalogId(orgName, catalogName string) resource.ImportStateIdFunc {
-	return func(s *terraform.State) (string, error) {
-		conn := testAccProvider.Meta().(*VCDClient)
-		org, err := conn.GetOrgByName(orgName)
-		if err != nil {
-			return "", err
-		}
-		catalog, err := org.GetCatalogByName(catalogName, false)
-		if err != nil {
-			return "", err
-		}
-		return orgName + "." + catalog.Catalog.ID, nil
-	}
-}
-
-// testAccFindValuesInSet finds several elements as belonging to the same item in a set
-// * resourceName is the complete identifier of the resource (such as vcfa_vapp_access_control.Name)
-// * prefix is the name of the set (e.g. "shared" in vApp access control)
-// * wanted is a map of values to check (such as {"subject_name" : "xxx", "access_level": "yyy"})
-// The function returns successfully if all the wanted elements are found within the same set ID
-// For example, given the following contents in the resource:
-//
-//	"shared.2503357709.access_level":"FullControl",
-//	"shared.3479897784.user_id":"urn:vcloud:user:ec571e04-7e75-4dc5-8f53-c3ef63b9b414",
-//	"shared.2503357709.user_id":"urn:vcloud:user:465308a5-7456-42c8-939c-bd971b0e0d3f",
-//	"shared.2503357709.subject_name":"ac-user1",
-//	"shared.3479897784.subject_name":"ac-user2",
-//	"shared.3479897784.access_level":"Change"
-//
-// We pass "shared" as prefix, and map[string]string{"subject_name": "ac-user1", "access_level": "FullControl"} as wanted
-// The function will match the elements belonging to set "2503357709", and return successfully, because both elements were found.
-func testAccFindValuesInSet(resourceName string, prefix string, wanted map[string]string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-
-		rs, ok := s.RootModule().Resources[resourceName]
-		if !ok {
-			return fmt.Errorf("not found: %s", resourceName)
-		}
-
-		var matches = make(map[string]int)
-		for key, value := range rs.Primary.Attributes {
-			keyList := strings.Split(key, ".")
-			if len(keyList) == 3 {
-				foundPrefix := keyList[0]
-				setID := keyList[1]
-				foundKey := keyList[2]
-				for wKey, wValue := range wanted {
-					if foundPrefix == prefix && foundKey == wKey {
-						if wValue == value {
-							_, ok := matches[setID]
-							if !ok {
-								matches[setID] = 0
-							}
-							matches[setID]++
-						}
-					}
-				}
-			}
-		}
-
-		for _, value := range matches {
-			if value == len(wanted) {
-				return nil
-			}
-		}
-		return fmt.Errorf("resource %s - %d matches found - wanted %d", resourceName, len(matches), len(wanted))
-	}
-}
-
-// skipOnEnvVariable takes a TestCheckFunc and skips it if the given environment variable was set with
-// an expected value
-func skipOnEnvVariable(envVar, envValue, notes string, f resource.TestCheckFunc) resource.TestCheckFunc {
-	if os.Getenv(envVar) == envValue {
-		fmt.Printf("### Check skipped at user request - Variable %s - reason: %s\n", envVar, notes)
-		return func(s *terraform.State) error {
-			return nil
-		}
-	}
-	return f
-}
-
-// skipNoConfiguration allows to skip a test if NSX-T configuration is missing
-func skipNoConfiguration(t *testing.T, paramsMap StringMap) {
-	for key, value := range paramsMap {
-		if value == "" {
-			t.Skipf("[%s] Missing test config: No %s specified", t.Name(), key)
-		}
-	}
-}
-
-func skipNoNsxtAlbConfiguration(t *testing.T) {
-	generalMessage := "Missing NSX-T ALB config: "
-
-	if testConfig.Nsxt.NsxtAlbControllerUrl == "" {
-		t.Skip(generalMessage + "URL")
-	}
-
-	if testConfig.Nsxt.NsxtAlbControllerUser == "" {
-		t.Skip(generalMessage + "User")
-	}
-
-	if testConfig.Nsxt.NsxtAlbControllerPassword == "" {
-		t.Skip(generalMessage + "Password")
-	}
-
-	if testConfig.Nsxt.NsxtAlbImportableCloud == "" {
-		t.Skip(generalMessage + "Importable Cloud")
-	}
-}
-
 func timeStamp() string {
 	now := time.Now()
 	return now.Format(time.RFC3339)
@@ -1458,7 +865,7 @@ func timeStamp() string {
 //     contains a pattern that matches the test name.
 //  6. If the flag -vcfa-re-run-failed is true, it will only run the tests that failed in the previous run
 func preTestChecks(t *testing.T) {
-	handlePartitioning(testConfig.Provider.VcdVersion, testConfig.Provider.Url, t)
+	handlePartitioning(testConfig.Provider.TmVersion, testConfig.Provider.Url, t)
 	// if the test runs without -vcfa-pre-post-checks, all post-checks will be skipped
 	if !vcfaPrePostChecks {
 		return
@@ -1619,40 +1026,6 @@ func addToTestRunList(testName, fileType string) error {
 	return w.Flush()
 }
 
-// noTestCredentials helps to check if a config file with credentials is actually provided. It helps to conditionally
-// ignore tests in such case
-func noTestCredentials() bool {
-	return testConfig.Provider.User == ""
-}
-
-// skipTestForVcfaExactVersion allows to skip tests for specific VCFA version
-// exactSkipVersion must match exact VCFA version (e.g. 10.2.2.17855680)
-func skipTestForVcfaExactVersion(t *testing.T, exactSkipVersion, skipReason string) {
-	vcdClient := createTemporaryVCFAConnection(false)
-
-	vcfaVersion, err := vcdClient.Client.GetVcdFullVersion()
-	if err != nil {
-		t.Fatalf("Could not determine VCFA version")
-	}
-
-	expectedVersion, err := semver.NewVersion(exactSkipVersion)
-	if err != nil {
-		t.Fatalf("could not process versions")
-	}
-	if vcfaVersion.Version.Equal(expectedVersion) {
-		t.Skipf("skipping test on VCFA version %s because %s", exactSkipVersion, skipReason)
-	}
-}
-
-func skipTestForServiceAccountAndApiToken(t *testing.T) {
-	if testConfig.Provider.ApiToken != "" {
-		t.Skipf("skipping test %s because API token does not support this functionality", t.Name())
-	}
-	if testConfig.Provider.ServiceAccountTokenFile != "" {
-		t.Skipf("skipping test %s because Service Accounts do not support this functionality", t.Name())
-	}
-}
-
 func getOrgProviderText(providerName, orgName string) string {
 	orgProviderTemplate := `
 provider "vcfa" {
@@ -1670,13 +1043,11 @@ provider "vcfa" {
 }`
 
 	data := StringMap{
-		"OrgUser":         testConfig.TestEnvBuild.OrgUser,
-		"OrgUserPassword": testConfig.TestEnvBuild.OrgUserPassword,
-		"VcfaUrl":         testConfig.Provider.Url,
-		"SysOrg":          orgName,
-		"Org":             orgName,
-		"ProviderName":    providerName,
-		"Alias":           "",
+		"VcfaUrl":      testConfig.Provider.Url,
+		"SysOrg":       orgName,
+		"Org":          orgName,
+		"ProviderName": providerName,
+		"Alias":        "",
 	}
 	unfilledTemplate := template.Must(template.New("getOrgProvider").Parse(orgProviderTemplate))
 	buf := &bytes.Buffer{}
@@ -1721,122 +1092,4 @@ provider "vcfa" {
 		return ""
 	}
 	return buf.String()
-}
-
-// logState will save the current Terraform state in the log
-func logState(label string) resource.TestCheckFunc {
-	line := fmt.Sprintf("# %s", strings.Repeat("-", 80))
-	return func(s *terraform.State) error {
-		util.Logger.Println(line)
-		util.Logger.Printf("# %s\n", label)
-		util.Logger.Println(line)
-		for resName, res := range s.RootModule().Resources {
-			util.Logger.Println("#", line)
-			util.Logger.Printf("# %s\n", resName)
-			util.Logger.Println("#", line)
-			util.Logger.Printf("%# v\n", pretty.Formatter(res))
-			util.Logger.Println("#", line)
-		}
-		return nil
-	}
-}
-
-// checkVersion checks a given version against a version constraint
-// for example:
-// version "37.2" ; constraint "< 36.1" ; returns false
-// version "37.2" ; constraint "> 36.0" ; returns true
-// The function returns false in case of errors, but it is chatty in such instances, because it means
-// an error in the test configuration that should be rectified
-func checkVersion(version, versionConstraint string) bool {
-	caller := util.CallFuncName()
-	if version == "" {
-		util.Logger.Printf("[INFO-%s] provided version is empty: can't be compared with constraints '%s'\n", caller, versionConstraint)
-		fmt.Printf("[INFO-%s] provided version is empty: can't be compared with constraints '%s'\n", caller, versionConstraint)
-		return false
-	}
-	checkVer, err := semver.NewVersion(version)
-	if err != nil {
-		util.Logger.Printf("[INFO-%s] error getting a semver object from version '%s': '%s'\n", caller, version, err)
-		fmt.Printf("[INFO-%s] error getting a semver object from version '%s': '%s'\n", caller, version, err)
-		return false
-	}
-	// Create a constraint to check against the provided version
-	constraints, err := semver.NewConstraint(versionConstraint)
-	if err != nil {
-		util.Logger.Printf("[INFO-%s] error getting a constraint object from versionConstraint '%s': '%s'\n", caller, versionConstraint, err)
-		fmt.Printf("[INFO-%s] error getting a constraint object from versionConstraint '%s': '%s'\n", caller, versionConstraint, err)
-		return false
-	}
-	// No screen message below this line: what follows is a legitimate boolean result
-	if constraints.Check(checkVer) {
-		util.Logger.Printf("[INFO-%s] version '%s' satisfies constraints '%s'\n", caller, checkVer, constraints)
-		return true
-	}
-
-	util.Logger.Printf("[TRACE-%s] version '%s' does not satisfy constraints '%s'\n", caller, checkVer, constraints)
-	return false
-}
-
-// testCheckCatalogAndItemsExist checks that a catalog exists, and optionally that it has as many items as expected
-// * checkItems defines whether we count the items or not
-// * expectedItems is the total number of catalog items (includes both vApp templates and media items)
-// * expectedTemplates is the number of vApp templates
-// expectedMedia is the number of Media
-func testCheckCatalogAndItemsExist(orgName, catalogName string, checkItems, minimalCheck bool, expectedItems, expectedTemplates, expectedMedia int) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		if testAccProvider == nil || testAccProvider.Meta() == nil {
-			return fmt.Errorf("testAccProvider is not initialized")
-		}
-		conn := testAccProvider.Meta().(*VCDClient)
-		catalog, err := conn.Client.GetAdminCatalogByName(orgName, catalogName)
-		if err != nil {
-			return fmt.Errorf("error retrieving catalog %s/%s: %s", orgName, catalogName, err)
-		}
-		if !checkItems {
-			return nil
-		}
-
-		if catalog.AdminCatalog.Tasks != nil {
-			err = catalog.WaitForTasks()
-			if err != nil {
-				return err
-			}
-		}
-
-		items, err := catalog.QueryCatalogItemList()
-		if err != nil {
-			return fmt.Errorf("error retrieving catalog item list: %s", err)
-		}
-		vappTemplates, err := catalog.QueryVappTemplateList()
-		if err != nil {
-			return fmt.Errorf("error retrieving vApp templates list: %s", err)
-		}
-		mediaItems, err := catalog.QueryMediaList()
-		if err != nil {
-			return fmt.Errorf("error retrieving media items list: %s", err)
-		}
-		if minimalCheck {
-			// With minimal check, we only test that at least one item has been downloaded, rather than all ones
-			if len(items) == 0 && expectedItems > 0 {
-				return fmt.Errorf("catalog '%s' -expected at least 1 item - found 0", catalogName)
-			}
-			if len(vappTemplates) == 0 && expectedTemplates > 0 {
-				return fmt.Errorf("catalog '%s' -expected at lest 1 vApp template - found 0", catalogName)
-			}
-			if len(mediaItems) == 0 && expectedMedia > 0 {
-				return fmt.Errorf("catalog '%s' -expected at least 1 media item - found 0", catalogName)
-			}
-		} else {
-			if len(items) != expectedItems {
-				return fmt.Errorf("catalog '%s' -expected %d items - found %d", catalogName, expectedItems, len(items))
-			}
-			if len(vappTemplates) != expectedTemplates {
-				return fmt.Errorf("catalog '%s' -expected %d vApp templates - found %d", catalogName, expectedTemplates, len(vappTemplates))
-			}
-			if len(mediaItems) != expectedMedia {
-				return fmt.Errorf("catalog '%s' -expected %d media items - found %d", catalogName, expectedMedia, len(mediaItems))
-			}
-		}
-		return nil
-	}
 }
