+++ vcfa/structure.go
@@ -1,0 +1,25 @@
+package vcfa
+
+import "os"
+
+// Checks if a file exists
+func fileExists(filename string) bool {
+	f, err := os.Stat(filename)
+	if os.IsNotExist(err) {
+		return false
+	}
+	fileMode := f.Mode()
+	return fileMode.IsRegular()
+}
+
+// contains returns true if `sliceToSearch` contains `searched`. Returns false otherwise.
+func contains(sliceToSearch []string, searched string) bool {
+	found := false
+	for _, idInSlice := range sliceToSearch {
+		if searched == idInSlice {
+			found = true
+			break
+		}
+	}
+	return found
+}
